Move response modifier to the headers middleware
